% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{break_segments}
\alias{break_segments}
\title{Break the tree into its linear segments (one integer vector per segment).}
\usage{
break_segments(parents)
}
\description{
Break the tree into its linear segments (one integer vector per segment).
}
