% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{classify_nodes}
\alias{classify_nodes}
\title{Classify nodes into roots (0), leaves (1), branch points (2) and slabs (3).}
\usage{
classify_nodes(parents)
}
\description{
Classify nodes into roots (0), leaves (1), branch points (2) and slabs (3).
}
