% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{mls_xform}
\alias{mls_xform}
\title{Apply a moving least squares transform to points}
\usage{
mls_xform(reg, xyz, n_cores = NULL)
}
\arguments{
\item{reg}{An \code{mls_transform}, from \code{\link[=mls_transform]{mls_transform()}}.}

\item{xyz}{An \verb{(N, 3)} matrix of coordinates. A data frame is also accepted -- if it has
x/y/z columns (in any case) those are used, whatever else it carries, otherwise the
first three columns -- as is a bare length-3 vector. \code{nat::xyzmatrix()} gives you this
for a neuron.}

\item{n_cores}{Number of threads. \code{NULL} (default) uses all cores.}
}
\value{
An \verb{(N, 3)} matrix of transformed coordinates.
}
\description{
Apply a moving least squares transform to points
}
\examples{
src <- rbind(c(0, 0, 0), c(10, 10, 10), c(100, 100, 100), c(80, 10, 30))
trg <- rbind(c(1, 15, 5), c(9, 18, 21), c(80, 99, 120), c(5, 10, 80))
mls <- mls_transform(src, trg)
mls_xform(mls, c(50, 50, 50))

}
\seealso{
\code{\link[=mls_transform]{mls_transform()}}, \code{\link[=mls_affine]{mls_affine()}}
}
