% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{nblast}
\alias{nblast}
\title{Forward NBLAST of every query neuron against every target neuron.}
\usage{
nblast(
  q_points,
  q_vects,
  t_points,
  t_vects,
  q_alphas,
  t_alphas,
  smat_values,
  dist_edges,
  dot_edges,
  normalize,
  limit_dist,
  n_cores,
  precision,
  progress
)
}
\description{
Returns an (n_query, n_target) score matrix.
}
