% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp.R
\name{tps_xform}
\alias{tps_xform}
\title{Apply a thin-plate spline to points}
\usage{
tps_xform(reg, xyz, n_cores = NULL)
}
\arguments{
\item{reg}{A \code{tps_transform}, from \code{\link[=tps_transform]{tps_transform()}}.}

\item{xyz}{An \verb{(N, 3)} matrix of coordinates. A data frame is also accepted -- if it has
x/y/z columns (in any case) those are used, whatever else it carries, otherwise the
first three columns -- as is a bare length-3 vector. \code{nat::xyzmatrix()} gives you this
for a neuron.}

\item{n_cores}{Number of threads. \code{NULL} (default) uses all cores.}
}
\value{
An \verb{(N, 3)} matrix of transformed coordinates.
}
\description{
Apply a thin-plate spline to points
}
\examples{
src <- rbind(c(0, 0, 0), c(10, 10, 10), c(100, 100, 100), c(80, 10, 30))
trg <- rbind(c(1, 15, 5), c(9, 18, 21), c(80, 99, 120), c(5, 10, 80))
tps <- tps_transform(src, trg)
tps_xform(tps, c(50, 50, 50))

\dontrun{
# transform a neuron
n <- nat::Cell07PNs[[1]]
nat::xyzmatrix(n) <- tps_xform(tps, nat::xyzmatrix(n))
}

}
\seealso{
\code{\link[=tps_transform]{tps_transform()}}, \code{\link[=tps_affine]{tps_affine()}}
}
