% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{child_to_parent_dists}
\alias{child_to_parent_dists}
\title{Calculate child -> parent distances.}
\usage{
child_to_parent_dists(parents, x, y, z)
}
\description{
Calculate child -> parent distances.
}
