% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{dist_to_root}
\alias{dist_to_root}
\title{Return path length from a single node to the root.}
\usage{
dist_to_root(parents, node)
}
\description{
Return path length from a single node to the root.
}
