% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{geodesic_distances}
\alias{geodesic_distances}
\title{Geodesic distances between nodes.}
\usage{
geodesic_distances(parents, sources, targets, weights, directed)
}
\description{
Geodesic distances between nodes.
}
