% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{node_indices}
\alias{node_indices}
\title{For each node ID in \code{parents} find its index in \code{nodes}.}
\usage{
node_indices(nodes, parents)
}
\description{
Importantly this is 0-indexed to match indexing in Rust.
Roots will have parent index -1.
}
