% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{smat_auto_limit}
\alias{smat_auto_limit}
\title{The \code{limit_dist="auto"} value for a scoring matrix.}
\usage{
smat_auto_limit(smat_values, dist_edges, dot_edges, use_alpha)
}
\description{
The \code{limit_dist="auto"} value for a scoring matrix.
}
