% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{synapse_flow_centrality}
\alias{synapse_flow_centrality}
\title{Synapse flow centrality for each node.}
\usage{
synapse_flow_centrality(parents, presynapses, postsynapses, mode)
}
\description{
\code{presynapses}/\code{postsynapses} give the number of pre-/post-synapses at each node.
\code{mode} is one of "centrifugal", "centripetal" or "sum".
}
