% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{synblast}
\alias{synblast}
\title{Forward syNBLAST of every query neuron against every target neuron.}
\usage{
synblast(
  q_points,
  q_types,
  t_points,
  t_types,
  smat_values,
  dist_edges,
  dot_edges,
  normalize,
  n_cores,
  precision,
  progress
)
}
\description{
Returns an (n_query, n_target) score matrix.
}
